requirements.txt
streamlit
psycopg2-binary
pandas
python-dotenv
flask
opencv-python
opencv-python-headless
google-cloud-vision
google-cloud-storage
requests